bake auditing into the operator
