+++ CHANGELOG.md
@@ -76,8 +76,9 @@ gates with no optional application-layer coupling in the critical path.
   Boundary Plane, invoked at L5 Actuator egress and before audit publishing. The raw audit vault is
   removed; all audit data now passes through Sentinel-moderated (`SentinelModerateRaw`) storage.
 * **Repository root is now the g8eo module** - `services/g8eo/` content promoted to root. `go.mod`,
-  `go.sum`, and all `internal/` packages now live at the repo root. `cmd/g8eo`, `cmd/chaos_tester`,
-  `cmd/uap-ping` are top-level `cmd/` entries.
+  `go.sum`, and all `internal/` packages now live at the repo root. `cmd/g8eo` is the sole
+  top-level `cmd/` entry. The auditor and chaos tester are now CLI subcommands (`./g8e auditor`,
+  `./g8e chaos`).
 * **`web_session_id` → `cli_session_id`** - Session ID field renamed across all logging, events, and
   generated constants to reflect CLI-first architecture.
 * **Cursor-based queries removed** - All cursor-based database query patterns eliminated in favor of
