+++ internal/contracts/docs_drift_test.go
@@ -109,7 +109,7 @@ func TestEventsDocsAlignment(t *testing.T) {
 		canonicalEvents[entry.Value] = true
 	}
 
-	// Note: events documentation is in docs/architecture/protocol.md
+	// Note: events documentation is in docs/architecture/g8e.md
 	// This test is a placeholder for verifying documented events match the registry.
 	// For now, we verify the protocol file is valid and contains events.
 	assert.NotEmpty(t, canonicalEvents, "events.json must contain event definitions")
